Functions/AvengersInfo/requirements.txt
azure-functions
azure-identity
azure-mgmt-compute
azure-mgmt-resource
kubernetes